requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
# psycopg==3.1.18
psycopg2-binary==2.9.9


